AudioEngine: reduced precision for 32bit Win only
The fix provided in #1679 reduced the precision in tick size comparisons within the AudioEngine for both 64 and 32bit Windows systems. AFAICS there is no way around it or at least not a simple solution.

However, with this patch the lower precision fix only affect Win 32bit system (for which it is relevant).

Fixes #1687