PatternEditor: use stacked view when PE is locked
In song mode with the pattern editor locked, the pattern editor always shows the last pattern in the current song column. After having the previous "stacked" pattern view (rendering the width of the longest playing pattern even if it exceeds the width of the current one and draw active notes of other patterns in faint colors) extended to virtual patterns as well, it feels only natural to use it for a locked pattern editor too. This way it displays all pattern information currently playing.