NPR: fix width update
the minimum width of the NotePropertiesRuler as set to the width of the first pattern encountered when opening Hydrogen. This caused all subsequent `resize()` calls to _not_ alter the width of the widget properly. In case of song mode without a locked pattern editor and a pattern playing which is smaller than the longest currently playing one, the bar of the playhead glitches. In DrumPatternEditor and PianoRollEditor it is shown till the very length of the pattern but in the NPR it exceed its length