keep locked PE always in sync
in case the Pattern Editor was locked the bottom-most pattern was only selected while playing and moving to an adjacent column. Now, it is also selected when activating the PE lock mode using button click and when relocating to another column while playing is not running. The user can still select different patterns in case playing is not rolling.