PE: used stacked display for virtual patterns in song mode
Previously, while in song mode and without the pattern editor being locked, virtual patterns were not rendered in the "stacked pattern view" within the Pattern editor. This was unintuitive because when having a virtual pattern selected and switching from song to pattern mode the stacked view was toggled on and off for no apparent reason.