SoundLibrary: add dk name to SoundLibraryInfo
previously, all .h2pattern files were loaded and a couple of their content was written into SoundLibraryInfo files. But not the name of the associated drumkit. Everytime SoundLibraryPanel::updateTree() was called all .h2pattern were opened _again_ to read just this property which is redundant and expensive. The drumkit name is now integrated in the SoundLibraryInfo class and loaded only once (and on SoundLibraryDatabase update)