load drumkit metadata via SoundLibrary
there were still some instances using `Drumkit::loadNameFrom` - which reads the kits name disk given the absolute path to the drumkit folder - _after_ `SoundLibraryDatabase` was already set up. The cached versions are now used instead and the `loadNameFrom` function could be dropped altogether.

I also spotted a bug in `Drumkit::loadLicenseFrom`. In case instrument kit A holds an instrument of kit B and B holds an instrument of A and infinite loop was created. To circumvent this the function still uses cached drumkits but does not load kits into the database if they do not exist yet