PreferencesDia: update buffersize and samplerate
In case of the JACK audio driver buffer size and sample rate are set by the server and not Hydrogen. When switching from a different driver to JACK, both values were previously not updated and might not represent the correct ones. Now, both sample rate and buffer size are updated when restarting the audio driver within the `PreferencesDialog`