do not render notes for instruments without samples
in case a note is played back for an instrument without any samples, it passes till the rendering methods of the `Sampler` and is discarded there after logging some error messages. It does not really do any harm. On the other hand, it is not an error to click click the name of a freshly created instrument in the pattern editor (which will trigger a preview note).

In this patch I prevent notes to be passed to the `Sampler` in case their associated instrument has no samples. It is a fast check but adds complexity as it was added at a lot of places. Ideally I would like to put it in just one place, namely `Sampler::noteOn`, but the lifecycle of `Note` is rather intricate. As soon as it enters `Sampler::noteOn` `Sampler` takes responsibility of deleting it as part of its process routines. Some functions, like `AudioEngine::processPlayNotes`, expect it to still exist after `Sampler::noteOn`. I do not want to check for the presence of samples in `Sampler::process` as it would be done not just one but many times per note. I also do not want to touch the lifecycle of `Note` right now was we are in between a beta and a release.