keep virtual patterns in sync
When adding or removing a pattern from a virtual one while playing (the virtual one), the new pattern was not properly added and played back. A new method `Hydrogen::updateVirtualPatterns()` was introduced to keep things in sync.

There is an edge-case in which the behavior might not be ideal. When in stacked pattern mode and removing a pattern from a virtual one, that pattern is still played back as we have no way of knowing whether it was started explicitly by the user or because it was once part of a virtual pattern. At least not without adding further complexity. I think we can leave it like this and it is not worth the extra complexity.