separate legacy drumkit validity check
The default choice of checking legacy kits in b2b8659 has been reverted. The CLI validation was introduced in order to help the Debian maintainer with their packaging pipeline. In case the validation fails, all drumkits shipped need to be upgraded in order to prevent unnecessary log messages and upgrade of drumkit provided by Debian and installed on system-level.

Instead, a separate --legacy-check option in the CLI and an additional parameter in the OSC command VALIDATE_DRUMKIT for toggling the legcay check was introduced