LCDSpinBox: expose type & fix precision
`LCDSpinBox` instances instantiated as part of a .ui file do all get the default `LCDSpinBox::Type::Int`. A `setType` method was introduce to cope for it.

Setting the spinbox type to `LCDSpinBox::Type::Int` does now also set the number of allowed decimals to `0`. This prevents the user from inputing digits after point what is exactly waht I would expect from a integer spinbox.

There was also an issue with the precision / text formatting of values with digits after point. If one e.g. sets the tempo in a BPM widget to 130.1 and spins up and down a number of time or reopens the dialog, the number might be shown as 130.09999999999999992. This turned out to be a `QDoubleSpinbox` issue. I have to admit I was a little bit lured in by the name of the class and assumed we can treat its values as doubles. But we can not. When setting value to 40.55555555555 it seems to get rounded internally. Well, I do not think it is a problem that we can not use double precision in the widget but it is strange nevertheless. I will stick to the default `QString( "%1" ).arg( value())` representation of the number from here on. Trying to tweaks it just do not seem to work properly