LCDSpinBox: fix is modified behavior
Previously, if not set otherwise in the constructor, a value change in a `LCDSpinBox` caused the current song in Hydrogen to be marked "modified". While this was good for the spin boxes in the main window it also applied for those constructed as part of an .ui file. But marking it modified by altering a value in the PreferenceDialog of BPM widget was never intended.

Instead, the is modified feature needs now to be explicitly activated and is only set for the main window boxes