SER: tweak TempoMarker representation
I found it a little bit hard to determine the column a `TempoMarker` does correspond. To ease it, I introduced an additional horizontal line right in the target column right above the BPM text. In addition, tick highlighting is now persistent when clicking on a column to open the BPM dialog and I fixed alignment issue found at insane higher scaling factors