SER: convert x <-> column & tweak TempoMarkers
Introduce `SongEditorRuler::columnToX()` and `SongEditorRuler::xToColumn()` to remove all the duplicates of calculating it in the code. (This is more or less a copy of a method pair in `SongEditor` but there is no common base for both clases yet).

Introduce `SongEditorRuler::calcTempoMarkerRect()` to dynamically calculate the rectangle containing the tempo marker text. It does now scale with the width of the text generated from the markers BPM.

Tempo markers are now left aligned with the column it is associated with. Otherwise, the first tempo markers would overflow when setting a larger float value, liek 130.42