TagWidget: fix dropped tag input
When opening a `SongEditorPanelTagWidget`, entering some text, and pressing RETURN right away (without hitting the arrow keys beforehand or clicking somewhere using the mouse) the input is discarded and no `Tag` is created.

This happened because when editing an item of the `QTableWidget` in the tag dialog a persistent editor is opened for the item. RETURN triggers `SongEditorPanelTagWidget::on_okBtn_clicked()` _before_ closing the persistent editor. This caused the previous implementation, which was more efficient since it just checked those items of the table for which `itemChanged` events were triggered, to miss the item edited last (closing the editor and emitting `itemChanged` happened after the okBtn routine). To circumvent this problem, we now manually close the persistent editor for the current item in the okBtn routine and compare all item texts with their original state.