Director: fit text
only small and concise song names and tag labels could be displayed whole in the previous implementation. All other did overflow to the left and right and were cut.

Now, font size is set to a default value (the one used previously) whenever a label or song name changed and reduced until the font fits the window width.