disentangle EVENT_METRONOME
Till now `EVENT_METRONOME` was triggered on relocations and when a metronome note (would be) queued in the `AudioEngine::m_songNoteQueue` at `AudioEngine::m_pQueuingPosition`. The signal is than used in two different ways: 1. to indicate a metronome note by flashing the `LED` below the metronome button, 2. to update the beat and bar information displayed in the `Director` (with the assumption that a metronome should sound on every bar).

Note that **beat** as discussed in here has a fixed length of 48 ticks.

This lead to a couple of issues
- metronome LED fires on relocations (even though no metronome note was played back)
- pausing transport at the right time could already trigger metronome LED while corresponding note was not played yet (as the lookahead between queuing and transport position has to be passed first)
- bar information in `Director` did lag behind. (as `EVENT_METRONOME` was fired at queuing position and Director shows position information at transport position))

I disentangled both use cases of the event and split it into `EVENT_METRONOME` and `EVENT_BBT_CHANGED`.

## `EVENT_METRONOME`

Is now fired in `AudioEngine::processPlayNotes()` every time a metronome note is passed to the `Sampler` and only evaluated by the metronome LED. -> LED only flashes if a metronome note is actually played.

## `EVENT_BBT`

`TransportPosition` class does now hold two more members `m_nBar` and `m_nBeat` - coarse grained versions of the transport position displayed in `Director`. The event is fired each time either or both of these variables do change in `AudioEngine::updateTransportPosition()` for the transport position. -> position displayed in `Director` is always in sync.

The **tick** part of BarBeatTick is not implemented yet as it is not used yet. But using these changes we could easy display this information in the `PlayerControl` instead of time (or in addition) at some point in the future).