i18n: support translations in core part
It does not feel right to expect future maintainers to edit a fille in core `core/Preferences/Shortcuts` and another one in the gui part to introduce a new shortcut just to have all translatable strings residing in the gui part. Instead, translatable strings are now also allowed in core.