Shortcuts: store keys as QKeySequence instead of ints
this allows to use platform and locale dependent standard keys shipped with Qt. But it also introduces more complexity as the default shortcut creation must not be done within the inialization of the `Preferences` class as Hydrogen would crash. In order to access standard keys, a QApplication instance must be present. Default creation is thus deferred till after the bootstrap. But I think this is worth it.