MidiAction: tweak next_bar and previous_bar
the two actions used to move one bar back or ahead using MIDI or OSC command or a global shortcut were adapted to better fit in with nowadays AudioEngine.

In song mode, in case loop mode is disabled, jumping one bar ahead does not take effect in the last column of the song. If its enabled, jumping both back and forth does use periodic boundary conditions and start over at the last column or beginning of the song respectively.

In pattern mode - which was not covered previously at all - transport moves back to the beginning of the pattern. In case of `next_bar` pending stacked pattern will be activated/deactivating as well (which gives the whole action a feeling of jumping to the "next" patterns)