DEBUG: add temp. debug output for pipeline
quite frequently, but not every time, our macOS pipeline gets stuck in `TransportTest::testSampleConsistency`. When looking at the code I do not see what could cause this issue (and I can not reproduce it locally). But we had a similar issue once were the thread of the `DiskWriterDriver` did not queued an event containing the final 100% progress. GUI itself does not care but unit test did fail because of it.

I added some temporary debug print statements that may point in the right direction once this problem shows itself within our pipeline again.