AudioEngine: fix false positives in humanization test
occassionally the last note in the humanization test pattern was moved to the very end of pattern due to the humanization and was thus not picked up and counted. This lead to failure of the Ubuntu building pipeline. The unit test now waits till no more notes are enqueued in the song note queue.