PortMidi: tweak event loop and error formatting
The current implementation used `Pm_Poll` to check whether there is data available. But PortMidi API doc is discouraging its usage https://portmidi.github.io/portmidi_docs/group__grp__io.html  and suggests to, instead, just use `Pm_Read` and check whether the length is 0. If so, there is no new data.

Checking for a length zero was already done. But in case the return value of `Pm_Read` is negative, an error did occur and the returned length should be interpreted as error code. This was not done yet.

I also introduced a helper function to get a human-readable version of the retrieved error code