fix null MIDI port handling
In case no input and/or output port/device was chosen in Preferences > MIDI System `Preferences::m_sMidiPortName` and `Preferences::m_sMidiOutputPortName` were both set to "None". This string reappear at several other locations and was now commonized using a single static getter function for the Preferences class.

In addition, both `CoreMidiDriver` and `PortMidiDriver` did check for the presence of port/device "None" and the latter did log a message complaining that it did not found it. These checks and logs are now suppressed.