Common MIDI message type inference
the particular type of a MIDI message can be determined by looking at the first 8bit of the message. Which value corresponds to which type is defined by the standard itself and is not dependent on a particular driver. Thus, we can use common code for our drivers to infer the type.

Using this patch `CoreMidiDriver`, `PortMidiDriver`, and `JackMidiDriver` do support all possible types of MIDI messages. Missing ones were added to the `MidiMessageType` enum.

fixes #1776