Fix tool tip of MIDI-learnable widgets
In 1.2 I introduced a tool tip for all `MidiLearnable` widgets showing what `Action` they do represent and which MIDI event they are mapped to.

But it was not the whole story yet. An Action can be mapped to several MIDI Events which all have to be listed in the tool tip.

In addition, those MIDI-learnable are just a more convenient way to add rows to the `MidiTable` in the `PreferencesDialog`. All mappings done their and all changes apply have to be represented in the tool tips as well. Previously, just those mappings done using `MidiSenseWidget` were used