ensure consistency of MidiMap
Within `MidiInput` all `Action`s in `MidiMap` mapped to an incoming MIDI event were first retrieved, then altered by assigning the MIDI events value, and finally send to the `MidiActionManager` to be handled.

Altering the actions stored in the `MidiMap` might be intended and efficient (at least at some point in time when it did not caused problems) but quite bad style and a place where there is likely to happen something bad in the future.

I introduced copy constructors for `Action` and assured that the actions within the MidiMap are not altered anymore