MidiMap: fix superfluous mappings
Since 1.2 an incoming MIDI event can be mapped to several Actions (the UX also suggest this to be possible previously. But only the last Action took precedence and all other were discarded).

But the check determining whether the MIDI event was already mapped to this Action was flawed and multiple instances of the same incoming message could be mapped to the same Action (causing it to be triggered multiple times).