PatternEditorPanel: fix layout
The layout of the `PatternEditorPanel` was kinda bricked since I reworked parts of the GUI. Its strings were translatable but one had to translate them with exactly the same number of characters. Else the text would overflow and look ugly. The Combo box for selecting the resolution of the grid was quite big since the we added the verbose versions of the resolutions and it was cut off when shrinking the main window.

With this fix the reolution combo box is not cut off but shrinked/expanded in such a way that it uses the available space. Since the layout does now use some actual QLayout instead of widget of hard-coded size at hard-coded positions, translations of the labels can be way more flexible and are not limited to the number of characters used in the English one. Also, the labels do use the available space more efficiently and are bigger when setting the font size to "larger".