PatternEditorPanel: fix resolution combo width
having the combo for choosing the resolution to use an expanding size policy was a left-over from an intermediate state.

It is way better to make is `QSizePolicy::Preferred`. This way the widget (and the surrounding panel part) is only as wide as the largest text in the combo box but still has all the flexibility and benefits in terms of resizing and translation