fix spurious modified flags on opened songs
Opening the initial song during startup reset the modified flag properly. But when opening another song within a running Hydrogen instance the freshly opened song is marked modified.

`CoreActionController::setSong` is now properly resetting the modified flag when setting (opening) a different song. In addition, updating the widgets in the `InstrumentEditor` to the values of the opened song, also introduced a spurious modified flag. This is now reset explicitly too