PortMidiDriver: create virtual device on "None"
when setting "None" for MIDI input or output PortMidi does now create a virtual device instead of none at all. This means when started there is no MIDI connection established. But external applications will be able discover Hydrogens ports and to connect to them. This PortMidi feature is only available on macOS and Linux but not on Windows