InputCaptureDialog: add title and fix type MIDI
when `Type::IntMidi` is selected the minimum and maximum possible values are set to 0 and 127 respectively (which is the overall reason this type was introduced).

For actions requiring multiple input captures the labels used for the individual captures, like "Instrument Number", did not hint which action was actually triggered. To fix this, an optional title label was introduced into the `InputCaptureDialog`.