cmake: print library versions
the code for creating summary of the status of each package in cmake did contain the `<IBRARY>_VERSION` variable. However, this one is nowhere set in the code and probably is an artifact from former versions of the cmake config script. If available - which is true for most packages -, we use the `pkg-config` support of cmake to find the path and version of a library. The later, however, is written to `PC_<LIBRARY>_VERSION`. All of those are now written to the summary printed during the build process.