cmake: use CXX flags from environment
instead of the `CMAKE_CXX_FLAGS` defined as an environmental variable to the `cmake` command, the internal but not yet set `CMAKE_CXX_FLAGS` variable is used when setting this very variabale. That is most probably a bug.

Flags can now be set in the environment and used will be used by `cmake`.