DWD: more debug messages
in https://ci.appveyor.com/project/mauser/hydrogen/builds/47546760/job/7wr7cciduryo86ij it seems Hydrogen stopped while processing the song and writing data.

It is not evident what is going wrong yet but currently I do see two possible causes: 1) for some reason the process callback is not able anymore to acquire the mutex of the audio engine (which would be really bad) or 2) the pthread started by the `DiskWriterDriver` suddenly dies. 2) is we should take care of regardless of whether it is the cause of the sporadic failures of the macOS pipeline or not.