AppVeyor/AppImage: fix discovery of libhydrogen.so
without pointing explicitly to the `libhydrogen-X.X.X.so` object in the installed AppDir using `LD_LIBRARY_PATH` or a RPATH the linker fails the resolve the shared object properly. This still worked on system with a libhydrogen already installed in /usr but not on clean ones. The `--library` option for `linuxdeploy` does only exchange libraries already found and does not affect their initial discovery