xsd: allow unbound number of notes in pattern
There was a limitation in the XSD file that set the maximum number of allow notes within a pattern to `1000`. This number can easily be reached by setting resolution to `off` and filling all notes for a couple of instruments. Curiously, Qt treated this number wrong and also deemed paterns containing more than `100` notes as invalid.

Once considered invalid, Hydrogen thinks this pattern is of a legacy format. But since it is not, loading partially fails.

I removed the constraint for a maximum number of notes.

fixes #1827