Jack: base support check on executables
since `libjack.so` is preinstalled on all Linux OSs I checked so far, looking it up makes not a good check for JACK support. Instead, a subprocess will now attempt to call
- `jackd`
- `jackdbus`
- `pw-jack`

and support is given in case any of these calls succeeded (just checking version or displaying help. Do not start anything).