h2map loading and display V1
improved version of .h2map loading. The drumkit map is now moved from the `SoundLibraryDatabase` into `Drumkit` and `Song` (for the current kit) and associated with a dedicated class `DrumkitMap`.

also, the drumkit properties dialog is now tab-based and the mappings can be configured in a dedicated tab. (just displayed for now)