Make ADSR tolerant of negative length notes
The AudioEngine could (and still can in rare circumstances) calculate
the length of a note as negative. Previously this would have caused
the ADSR implementation to write past the beginning of the buffer
as pre-Release states would assume that Release must be in the
future.

Instead, detect an early Release point and move directly to the
Release state.