Sampler: account for note length during rendering
it seems the custom note length defined by the user never made it into the `Sampler`. At least properly.

Previously, the remaining frames of a samples were always calculated from the original length of the sample. Not the user defined one. The original length was also the one added to the sample position of the selected layer being rendered. At the end of the user defined length this caused the sample position of the selected layer to be _beyond_ the note length and the note end provided to `ADSR::applyADSR()` to be negative. The latter than ended rendering.