MainForm: fix Drumkit Save As
whenever the last loaded kit is not present on the current system.

First, the path of the last loaded kit found in a .h2song is tried. Second, the last loaded kit name is used. If none of these resulted in a valid drumkit, an empty one will be used as basis instead.