xsd: add sampleEditor props to drumkit.xsd
Whenever a sample was adjusted this adjustment just lived in the song (.h2song) but was a) always discarded when loading a different drumkit or reloading the current one and b) could not be stored to a drumkit. The only (and undocumented) way to really use it was to make copies of the song and work with them over actual drumkits.

This maybe was a good idea while the SampleEditor was still in development. But there has not been much work in the past 10/15 years. And it is still somewhat poorly supported.

By moving all the properties it writes to disk into the kit itself we can now make settings of the SampleEditor portable between different setting and allow users to work with it in a way more intuitive way