Hydrogen: create SLDB before song
and use empty song instead of plain one. The latter could cause a crash in the `AudioEngine::prepare` since `Song::m_pPatternList` is initialized with a nullptr. The empty song, on the other hand, is a fully fledged one and can be used right away.

The `SoundLibraryDatabase` does not require the `Hydrogen` singleton and could be created prior to its initialization. But I think creating the DB and passing it as a pointer to other constructors depending on it is the most cleanest and easiest to grasp way to structure the startup process.