cmake: add internal DISPLAY_VERSION_PIPELINE var
using which the construction of a DISPLAY_VERSION within CMake can be bypassed. This is added as an auxiliary measure so the pipelines will always know the name of the created (Windows) artifacts and both the artifacts name and the version displayed inside do match.

Nevertheless, the TARGET_VERSION in .appyevor and major, minor, patch in CMakeLists.txt must still match.