tests: .h2song backward compatibility tests
add songs created with various versions of Hydrogen >1.0.0 (versions which I was able to compile locally without altering my setup) and check whether they a) can be loaded and b) if there are no missing samples.

All songs feature instruments from different kits as well as new instruments with freshly loaded samples. Newer ones also have some SampleEditor settings.

In general, less files would be required as song format did not change in every single on of them with respect to the instrument list of the song. But, on the other hand, this is definitely a nice-to-have for the unit tests