support relative sample paths in .h2song
In case a sample file is stored using a relative path containing a directory separator, like `../sample.flac`, within a `.h2song`, it is now read and resolved by Hydrogen when loading the song.

This was introduced in order to provide more sophisticated song loading tests.

However, it is not officially announced or documented as - for now - Hydrogen would overwrite the path using an absolute one when saving the song again and does not take care of moving the assets when "save as" the song to a different location.

Also note that a separator is strictly required. `sample.flac` is still interpreted as a sample from the corresponding drumkit