reintroduce old drumkit mapping
this way notes can still be used after switching to a different kit on `master` once this PR is merged (and before implementing the new, more sophisticated drumkit switching).

But it is not 100% behaving like the former version. Previously, the instrument ids of the new DK were discarded and reset to 0..N in their order of appearance. For kits with IDs set in a non-default way, there might be a little strange behavior.