SoundLibrary: add Duplicate DK action
Instead of requiring to load a kit first before being able to store it using a different name/to duplicate it, the menu opening up when right-clicking the `SoundLibraryPanel` now features a "Duplicate" action for all drumkits.

In addition, the window title for the `SoundLibraryPropertiesDialog` does now clearly indicate whether we are altering properties or creating a new drumkit.