move code component handling into core
whenever an instrument is added or delete, Hydrogen has to take care that associated `InstrumentComponent`s match the `DrumkitComponent`s. This code was, I do not know why, located almost exclusively in src/gui. I moved it into the `Drumkit` class.

UX change: adding and deleting drumkit components does now work using a heuristic. When adding all new components will be introduced into the drumkit as well. When removed all components the instruments has samples in but no other instrument has will be removed. This allows for having empty components in the drumkit (e.g. when freshly adding one in the GUI) but might lead to cluttering the kit with empty components