load drumkit after importing
it via MainForm > Drumkit > Import. It replaces the current one.

When importing via the SoundLibrary, the kit will not be loaded. Instead, it appearing in the widget under mouse is enough feedback. This also is supposed to highlight the distinctiong between the kit of the current song and the kits in the SoundLibrary.