SLP: drop highlighting of drumkit items
Previously, the drumkit the currently selected instrument is associated with was highlighted using a darker background color in the `SoundLibraryPanel`. I dropped this features as this intended meaning is not evident from UI/UX. It looks more like the drumkit used was highlighted. But this is wrong and gets even worse with the introduction of proper song kits in 1.3.0.

I also toyed with highlighting corresponding instrument item instead of the drumkit one but this is hidden most of the time and too much rendering for a hidden feature. It also bears the problem that an instrument can easily be renamed or all its layers can be replaced.

Instead, associating an imported instrument with a drumkit will be done by a tooltip in the `PatternEditorInstrumentList`.