drop WidgetWithHighlightedList
previously, this base class allowed for highlighting the row in the pattern list of the song editor and the instrument list in the pattern editor when opening a dialog window using right click. Before overhauling the UI in 1.2.0 this was important since it was not evident at all which line was clicked / the dialog corresponds to.

But now theming clearly indicates the selected row and all this code can be dropped.