SLP: forbid deleting drumkit if samples contained
Previously, deleting a drumkit was only prevented in case it was the last one loaded. But this is quite dangerous. It was possible that an instrument was loaded from another kit (by drag & drop) and this very kit could be deleted. -> nullptr -> bad stuff.

Now, `SoundLibraryPanel` checks whether there is any sample used and loaded in the current song belonging to the kit in question. If anything is found, a critical error is reported.