MainForm: undo add component using SE_switchDrumkitAction
adding and deleting a drumkit component is a quite complex task. Deleting a fully-fledged one also deletes all corresponding `InstrumentComponent`s including their settings, layers, and samples.

In order for undoing/redoing such a task to work as cleanly as possible we create an entire new kit and use the facilities for drumkit switching. This comes with an memory overhead as we store the whole drumkit and not just a single component. But since these include no sample data it should be ok.